requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
python-dotenv
pandas
Pillow
numpy
matplotlib